introduce UnregisteredContentTypeAbstract (WP-782)
